Split Components into separate mod
