+++ src/pipeline/mod.rs
@@ -1,3 +1,4 @@
+pub mod components;
 pub mod config;
 pub mod deferred;
 pub mod glow;
@@ -5,7 +6,6 @@ pub mod render_pass;
 pub mod shaders;
 pub mod shadow;
 
-pub use config::Config;
 pub use render_pass::{CompositionPassComponent, RenderPassComponent, ScenePassComponent};
 
 use coarse_prof::profile;
@@ -14,258 +14,14 @@ use log::info;
 use glium::{uniform, Surface};
 
 use crate::fxaa::{self, FXAA};
-use crate::shader::{ToUniforms, ToVertex, UniformInput};
+use crate::shader::ToUniforms;
 use crate::{
-    object, scene, screen_quad, shader, Context, DrawError, Instancing, Light, RenderList,
-    RenderLists, Resources, ScreenQuad,
+    object, scene, shader, Context, DrawError, Instancing, Light, RenderLists, Resources,
+    ScreenQuad,
 };
 
-use deferred::DeferredShading;
-use glow::Glow;
-use shadow::ShadowMapping;
-
-struct Components {
-    shadow_mapping: Option<ShadowMapping>,
-    deferred_shading: Option<DeferredShading>,
-    glow: Option<Glow>,
-}
-
-#[derive(Debug, Clone)]
-struct ScenePassSetup {
-    shadow: bool,
-    glow: bool,
-}
-
-struct ScenePass<I: ToVertex, V> {
-    setup: ScenePassSetup,
-
-    /// Currently just used as a phantom.
-    #[allow(dead_code)]
-    shader_core: shader::Core<Context, I, V>,
-
-    program: glium::Program,
-
-    instancing: Instancing<I>,
-}
-
-impl Components {
-    fn create<F: glium::backend::Facade>(
-        facade: &F,
-        config: &Config,
-        target_size: (u32, u32),
-    ) -> Result<Self, CreationError> {
-        let shadow_mapping = config
-            .shadow_mapping
-            .as_ref()
-            .map(|config| ShadowMapping::create(facade, config))
-            .transpose()
-            .map_err(CreationError::ShadowMapping)?;
-
-        let deferred_shading = config
-            .deferred_shading
-            .as_ref()
-            .map(|config| {
-                DeferredShading::create(facade, &config, shadow_mapping.is_some(), target_size)
-            })
-            .transpose()
-            .map_err(CreationError::DeferredShading)?;
-
-        let glow = config
-            .glow
-            .as_ref()
-            .map(|config| Glow::create(facade, config, target_size))
-            .transpose()
-            .map_err(CreationError::Glow)?;
-
-        Ok(Self {
-            shadow_mapping,
-            deferred_shading,
-            glow,
-        })
-    }
-
-    fn create_scene_pass<F, I: ToVertex, V>(
-        &self,
-        facade: &F,
-        setup: ScenePassSetup,
-        mut shader_core: shader::Core<Context, I, V>,
-    ) -> Result<ScenePass<I, V>, crate::CreationError>
-    where
-        F: glium::backend::Facade,
-        I: UniformInput + Clone,
-        V: glium::vertex::Vertex,
-    {
-        info!(
-            "Creating scene pass for I={}, V={}",
-            std::any::type_name::<I>(),
-            std::any::type_name::<V>(),
-        );
-
-        if let Some(glow) = self.glow.as_ref() {
-            if setup.glow {
-                shader_core = ScenePassComponent::core_transform(glow, shader_core);
-            } else {
-                // Whoopsie there goes the abstraction, heh. All good though.
-                shader_core = glow::shaders::no_glow_map_core_transform(shader_core);
-            }
-        }
-
-        if let Some(shadow_mapping) = self.shadow_mapping.as_ref() {
-            if setup.shadow {
-                shader_core = ScenePassComponent::core_transform(shadow_mapping, shader_core);
-            }
-        }
-
-        if let Some(deferred_shading) = self.deferred_shading.as_ref() {
-            shader_core = ScenePassComponent::core_transform(deferred_shading, shader_core);
-        } else {
-            shader_core = shaders::diffuse_scene_core_transform(shader_core);
-        }
-
-        let program = shader_core.build_program(facade, shader::InstancingMode::Vertex)?;
-
-        let instancing = Instancing::create(facade)?;
-
-        Ok(ScenePass {
-            setup,
-            shader_core,
-            program,
-            instancing,
-        })
-    }
-
-    fn composition_core(&self, config: &Config) -> shader::Core<(), (), screen_quad::Vertex> {
-        let mut shader_core = shaders::composition_core();
-
-        if let Some(deferred_shading) = self.deferred_shading.as_ref() {
-            shader_core = CompositionPassComponent::core_transform(deferred_shading, shader_core);
-        }
-
-        if let Some(glow) = self.glow.as_ref() {
-            shader_core = CompositionPassComponent::core_transform(glow, shader_core);
-        }
-
-        if let Some(_) = config.hdr {
-            // TODO: Use factor
-            shader_core = shaders::hdr_composition_core_transform(shader_core);
-        }
-
-        if let Some(gamma) = config.gamma_correction {
-            shader_core = shaders::gamma_correction_composition_core_transform(shader_core, gamma);
-        }
-
-        shader_core
-    }
-
-    fn clear_buffers<F: glium::backend::Facade>(&self, facade: &F) -> Result<(), DrawError> {
-        self.shadow_mapping
-            .as_ref()
-            .map(|c| c.clear_buffers(facade))
-            .transpose()?;
-        self.deferred_shading
-            .as_ref()
-            .map(|c| c.clear_buffers(facade))
-            .transpose()?;
-        self.glow
-            .as_ref()
-            .map(|c| c.clear_buffers(facade))
-            .transpose()?;
-
-        Ok(())
-    }
-
-    fn scene_output_textures(
-        &self,
-        setup: &ScenePassSetup,
-    ) -> Vec<(&'static str, &glium::texture::Texture2d)> {
-        let mut textures = Vec::new();
-
-        textures.extend(
-            self.deferred_shading
-                .as_ref()
-                .map_or(Vec::new(), ScenePassComponent::output_textures),
-        );
-
-        if setup.glow {
-            textures.extend(
-                self.glow
-                    .as_ref()
-                    .map_or(Vec::new(), ScenePassComponent::output_textures),
-            );
-        }
-
-        textures
-    }
-
-    fn scene_pass_to_surface<I, V, S>(
-        &self,
-        resources: &Resources,
-        context: &Context,
-        pass: &ScenePass<I, V>,
-        _render_list: &RenderList<I>,
-        target: &mut S,
-    ) -> Result<(), DrawError>
-    where
-        I: ToUniforms + ToVertex,
-        V: glium::vertex::Vertex,
-        S: glium::Surface,
-    {
-        let params = glium::DrawParameters {
-            backface_culling: glium::draw_parameters::BackfaceCullingMode::CullClockwise,
-            depth: glium::Depth {
-                test: glium::DepthTest::IfLessOrEqual,
-                write: true,
-                ..Default::default()
-            },
-            line_width: Some(2.0),
-            ..Default::default()
-        };
-
-        let uniforms = (
-            context,
-            self.shadow_mapping
-                .as_ref()
-                .map(|c| c.scene_pass_uniforms(context)),
-        );
-
-        pass.instancing.draw(
-            resources,
-            &pass.program,
-            &uniforms.to_uniforms(),
-            &params,
-            target,
-        )?;
-
-        Ok(())
-    }
-
-    fn scene_pass<F, I, V>(
-        &self,
-        facade: &F,
-        resources: &Resources,
-        context: &Context,
-        pass: &ScenePass<I, V>,
-        render_list: &RenderList<I>,
-        color_texture: &glium::texture::Texture2d,
-        depth_texture: &glium::texture::DepthTexture2d,
-    ) -> Result<(), DrawError>
-    where
-        F: glium::backend::Facade,
-        I: ToUniforms + ToVertex,
-        V: glium::vertex::Vertex,
-    {
-        let mut output_textures = self.scene_output_textures(&pass.setup);
-        output_textures.push((shader::defs::F_COLOR, color_texture));
-
-        let mut framebuffer = glium::framebuffer::MultiOutputFrameBuffer::with_depth_buffer(
-            facade,
-            output_textures.into_iter(),
-            depth_texture,
-        )?;
-
-        self.scene_pass_to_surface(resources, context, pass, render_list, &mut framebuffer)
-    }
-}
+use components::{Components, ScenePass, ScenePassSetup};
+use config::Config;
 
 pub struct Pipeline {
     components: Components,
@@ -561,16 +317,7 @@ impl Pipeline {
         facade: &F,
         target_size: (u32, u32),
     ) -> Result<(), CreationError> {
-        if let Some(deferred_shading) = self.components.deferred_shading.as_mut() {
-            deferred_shading
-                .on_target_resize(facade, target_size)
-                .map_err(CreationError::DeferredShading)?;
-        }
-
-        if let Some(glow) = self.components.glow.as_mut() {
-            glow.on_target_resize(facade, target_size)
-                .map_err(CreationError::Glow)?;
-        }
+        self.components.on_target_resize(facade, target_size)?;
 
         self.scene_color_texture = Self::create_color_texture(facade, target_size)?;
         self.scene_depth_texture = Self::create_depth_texture(facade, target_size)?;
@@ -610,15 +357,19 @@ impl Pipeline {
 
 #[derive(Debug)]
 pub enum CreationError {
-    ShadowMapping(shadow::CreationError),
-    DeferredShading(deferred::CreationError),
-    Glow(glow::CreationError),
     FXAA(fxaa::CreationError),
+    Components(components::CreationError),
     CreationError(crate::CreationError),
 }
 
 impl From<crate::CreationError> for CreationError {
     fn from(err: crate::CreationError) -> CreationError {
         CreationError::CreationError(err)
+    }
+}
+
+impl From<components::CreationError> for CreationError {
+    fn from(err: components::CreationError) -> CreationError {
+        CreationError::Components(err)
     }
 }
